airflow_base/requirements.txt
apache-airflow-providers-docker
python-dotenv
docker